integrate wu palmer calculation in wu palmer plugin, delete wu palmer cache generator plugin
